+++ CHANGELOG.md
@@ -1,18 +1,33 @@
+[1.0.5] - 2025-01-08
+
+Fixed
+
+* Fixed the issue where only files under the project path could be selected
+
+Changed
+
+* Changed the file chooser to use the project resouces path as the default path
+
+
 [1.0.4] - 2025-01-07
 
 Added
+
 * Added a button to select files or directories for YAML processing
 * Integrated IntelliJ IDEA's file chooser for selecting files or directories
 
 Changed
+
 * Reorganized buttons in the UI for better usability
 * Renamed buttons to more appropriate names
 
 [1.0.3] - 2024-12-26
 
 Added
+
 * Added automatic encryption/decryption for YAML files
 * Support for multi-module projects
 
 Fixed
+
 * Fixed handling of environment variables in jasypt configuration